✨ Add `hash_value` to `hashed_string_view`
The `hashed_string_view` type never implemented either `std::hash` or
the newly added `hash_value` function, which makes it unusable in
hashed containers without a custom comparator.

This just adds the missing pieces to make it easier to use. In the
process, the file has been modernized as well.