🔨 Update `event`'s type-erasure technique
The technique used for `event`'s type-erasure was previously a little
janky in that the handler's return value was conditionally used to
represent either the priority of the id. Additionally, a lot of
`const_cast`'s were required, along with possibly optional inputs.

This updates the technique to instead only communicate via a `payload`
object which is a `void*`-erased type that can be any of the specified
payloads. This is semantically far clearer to read and to update, since
introducing a new operation only requires introducing a new payload type
and a new handler in the process.

All tests pass from this change, which indicate no regression in
behavior.

While implementing this, the C++ style was updated to reflect the modern
formatting guidelines.