🔨 Use absolute paths in `modernize-file` script
The `modernize-file` script for `clang-tidy` was not working properly
when used from a submodule, in-part due to the relative paths not
being seen as the same path as the ones specified in the compilation
database. Since the database uses absolute paths, this has been changed
so that all inputs will also use absolute paths, which appears to help
the parsing out.