✏ Rename tool scripts
The various tool scripts were previously named in a way that was not
particularly obvious or clear. In particular, `bootstrap` was used to
configure the project, `copyright-verifier` was used to check the year
of the copyright, etc.

This updates the tools to use better and more generic terms to be a
little more clear for the intentions.