🧹 Convert `ctor_function` to nebloid
The `ctor_function` previously required the user to construct it
manually with `{}` in order to pass it into functions that accept
callbacks. This is a bit obtuse in terms of design, since the object is
otherwise empty -- meaning the construction is largely just aesthetically
displeasing.

To compensate for this, the `ctor_function` has been converted to a
"Nebloid", which is a functor object with a `inline constexpr` constant.
This allows `ctor_function<T>` to be sufficient as the argument to
functional systems.