✨ Glob only for git-tracked sources in tools
The various python tools were previously globbing for all sources in the
repository that matched the glob expression. However, this was too
loose of a process, and would result in sources found under the build
directories to be (incorrectly) detected. This includes sources that
may be introduced via CPM, generated sources, etc.

To fix this, a new function to `git` was added for detecting which files
are tracked by git, and then these files are filtered from the
SourceGroup inputs when finding `all`.