🧹 Clean up reporting output
The reporting output previously cleared the progress bar before each
redraw -- resulting in a flicker. It also output the file list in the
order the files were processed, which -- when multithreaded -- changes
upon subsequent invocations.

To address these things, the bar is no longer cleared first to prevent
flickering, and the file output is always sorted before being used.