🧹 Disable clang-format pre-commit hook
As nice as `clang-format` would be in practice for this repository, the
custom styling done consistently throughout the repository really does
not play nicely with `clang-format`. Until there is some understanding
on a proper formatting style, this will have to be disabled to avoid
C++ files from being shot-down.

The unfortunately thing with `clang-format` is that it does not play
nice with trailing-return-type functions. It pushes function names to
the next line first rather than return-types, which defeats the
purpose of a unified `auto name -> result` syntax.