✨ Enhance pre-commit output
The pre-commit output previously would hang before outputting the
result of a hook check. The downside with this approach is that long-
running checks like `clang-tidy` would result in it appearing to hang
until finally determining whether the check passed or failed.

This has been updated by printing out the current check/context first
and updating it after the status has been determined.

In order to do this properly, the column-length is now automatically
determined by the current terminal size to allow for the redraw to work
correctly. Otherwise something drawing over the terminal column-width
may result in two lines, since the first set of lines won't be
overwritten.