✨ Add git-handling for python
By default, git does not understand certain files written in python,
such as the git config hooks. Additionally, it won't prevent generated
files from being able to be added to the repository, such as the
`__pycache__` that may be created by having a python package.

This adds some handling into Git to support this, both in terms of
updating the gitignore file to hide `__pycache__`, but also in terms of
a `gitattributes` file for automatically determining the correct type
of file to use for diffing contexts. This also now includes content for
C++, even though this *should* normally be discovered automatically.