🩹 Fix CMake errors for Ubuntu
A couple failures occur when building for Ubuntu that don't occur for
Windows. The first failure is that CMake has difficulty determining
which version of OpenGL to search for -- Legacy, or modern.
The second is that the debian SDL2 package suffers from a bug where
the include of SDL-internal files are including from a path not
discoverable in the include-path. This problem appears to mostly
affect clang-builds, but still needs to be fixed.

This corrects those changes by updating the CMake to account for them.