🧹 Clean up `configure.py` script
The configuration script for generating the project in CMake has been
in need of a clean-up for a little while. In particular, the settings
were limited such that it was impossible to configure the project
without enabling tests, extras, and examples.

This changes the structure now so that they must be toggled *on*, but
may also be collectively enabled with a simple `--all` parameter.

During the process of this change, argument parsing output has been
cleaned up to group arguments, and provide appropriate/clean metavar
syntax to each one.