✨ Add views for associative containers
C++17 introduced the notion of a `string_view` type; an object that does
not *own* a string, but can still operate on an object that points to
one. This is a powerful pattern as it allows type-erasure from
string-like types like `const char*`, `char[]`, `std::string`, etc --
all while working in a single homogeneous non-polymorphic way.

C++20 takes this further with `span` to allow for any contiguous view of
data; but nothing exists to support associative or set-like data.
This introduces views for exactly that purpose. They are non-owning,
immutable objects that view, but do not mutate, existing data.

The `map_view` was added to view any associative data like `std::map` or
`std::unordered_map`. Similarly, a `set_view` has been added that views
any `std::set` or `std::unordered_set`, or the keys of a map-like type.