✨ Add `object_storage` utility
Working with raw memory in C++ can be complicated, especially when
producing contiguous sequences of elements that don't yet have
lifetimes. This often requires performing arithmetic to compute the
alignment of the objects, the size of the contiguous sequence of N
objects of M bytes each, etc.

To workaround this, a utility has been added that performs this logic
for us instead. This produces a buffer of N bytes per object that may
or may not contain a lifetime. It's already suitably aligned for the
object, meaning that offsetting into sequences is as simple as it
is for any array of normal objects.