✨ Add `thread_group` utility
It is frequently useful to be able to coordinate the joining or
detaching of several threads at once. Often, this requires constructing
a sequence, such as a `vector` or an `array`, and manually looping over
all threads to perform this.

The `thread_group` utility just provides a simple means of
encapsulating the group functionality on a sequence of threads.