🩹 Fix `commit-msg` hook requirements
Some of the commit-msg hook requirements were a little too strict, and
required that all messages contain a verbose description -- even when
it was not practical to do so. This was a bug caused by the
comment-lines being parsed in the hook when they should have been
ignored.

This has now been fixed.