🔨 Change int comparisons to functors
The `int_utilities` class contains functions for performing proper
sign-aware comparisons. By default, this deduces the type of both the
left and the right argument; however, this API is a little
cumbersome to use when trying to pass the comparator as an actual
`comparator`-style argument.

To compensate for this, the various comparators have been changed to
functor objects that are instantiated and kept as constants in the
`int_utilities` class. This allows for it to be used in a transparent
way in classes that are capable of supporting it, such as `std::map`
as a transparent functor object.