🔨 Add CMakePresets file
Modern versions of CMake now support "preset" functionalities which
bucket together compiler, toolchains, and configuration values into
a single JSON file. This takes the burden off of the builder from
needing to supply this information manual, and off of the author from
needing to supply wrapping scripts to simplify this.

This adds such a script into the repo for selecting one of the typical
modern compilers: clang/clang-cl, gcc, or msvc.