🧹 Fix use of doxygen '\c' tag
The '\c' tag in doxygen is meant to represent a 'code' tag, but
is implemented as an alias of the `\p` tag. The original documentation
uses `\c` as a code block, but this triggers doxygen warnings/errors.

This fixes all the warnings/errors by usijng backticks instead, which
provides the same markup.