🩹 Fix broken UTF conversion utilities
Several of the UTF conversion utilities were mysteriously checked-in in
a broken state. In particular, the tests used an incorrect number of
output characters. Latin1 was also broken, as the tests didn't account
for the fact that certain 2-byte UTF-8 codepoints and 1-character UTF-16
codepoints could properly translate to Latin1.

This fixes those bugs, and also updates a few other minor buggy areas,
such as empty-input detection (which failed for UTF-8, but succeeded
for most other conversion utilities).