✨ Add patching of compile-commands
CMakePresets introduce an interesting mechanic that allows bundling
features of compile-flags into a json file. This is extremely useful
for specifying build-paths, configure-time properties, etc; but it
complicates IDE integration for the compile_commands.json.
IDEs like vscode like to have a common/consistent place for the compile
commands, such as under `build/compile_commands.json` -- which specifying
specific build paths conflicts with (since there may be multiple).

This adds a CMake module that simply symlinks the top-level build
dir to the subdirectory if and only if the build directory is under
`build`.