📇 Change license for tests to public domain
Tests within this project were previously licensed under the MIT
license, along with the rest of the code. However, this restriction is
unnecessary for tests. If a user wants to reuse a test suite for their
own project to validate their code, it seems silly to require license
attribution.

This updates the test licenses to now be public domain, enabling full
reuse.