🩹 Fix `angle::constrained()` implementation
The `angle::constrained()` implementation was inherently broken in the
definition, but never caught because the template was never directly
instantiated. The issue is that the `angle` being computed is marked
`const`, but may be mutated if the angle exceeds a revolution.

This fixes the error for now; but unit tests should be added later to
ensure that this code is 100% correct.