🔨 Promote `unsafe_mesh_builder` to public API
Internal to the various `mesh_builder`s is an `unsafe_mesh_builder` API
that forces an explicit requirement on the order of function calls.
This is inherently unsafe, and thus not ideal for the public API;
however there are cases where mesh builders may not be part of the
general `render` library but may still be desirable to be used.

One such example is an OBJ loader, which may be used for tooling so that
the raw object file format can be converted into a more typical binary
serialized representation.