🎯 Add tests for `basic_angle`
The `basic_angle` class was previously untested, which lead to some
compilation failures in a function of the primary template. This adds
some basic tests to ensure that such failures will not occur in
practice, and to ensure that the code is actually valid.

Additionally, a mutating `constrain()` function was added to compliment
the non-mutating `constrained()` function.