🩹 Fix enum_traits `is_contiguous_enum`
The `enum_traits` implementation incorrectly specified the default value
of `is_contigous_enum` to be `true_type`, when it should be `false_type`.

The intended behavior is that it should be opt-in for devs, rather than
opt-out. Somewhere along the way, wires were crossed here and the
default was set incorrectly. This fixes the issue.