✨ Add `common_reference` implementation
C++20 introduces the trait `std::common_reference` which behaves similar
to `std::common_type` except it can produce reference types if both
inputs are also references. This is extremely useful for several
meta types and algorithms that want to propagate base-type references
back to the caller.

In order to implement this, `copy_cv` and `copy_cvref` were needed,
which are also added here.