✨ Add `compare_reduce`, `max`, and `min`
`std::min` and `std::max` are both limited and error-prone utilities,
for a few reasons:

1. The types must be homogeneous, which means comparison of unlike
   or similar types will force coercion
2. It returns a reference to the input, which may be dangling if the
   inputs were PR values, and
3. It is not variadic and can only work with 2 inputs

This implements a safer functional object for generic compare reduction
which uses `common_reference` for the return type to ensure no dangling
references are possible, and so that heterogeneous variadic inputs
can be provided.

In the process, a set of new transparent comparators have also been
added that safely allow comparison across integer values of differing
signs. This is the default for the compar reduction.