🔨 Return on empty UTF8 encoded range
The utf8_encoding type was not guarding against cases where the range
was empty, which resulted in potential errors occurring during decoding
if the range was truncated, or if the string being decoded was empty.

This adds the appropriate guard. Additionally, an extra `ALLOY_UNLIKELY`
attribute was marked around the UTF16 encoding guard, as these cases
should be rare in practice and not the optimized path.