✨ Add trait for `always_false`
When working with `static_assert` in C++, it's illegal for the condition
to always produce a `false` state -- leading to compile-time failures
even if the condition exists in a template.

A workaround for it is to use a template-type that always evaluates to
false, since this requires a dependent-template. This commit introduces
such a workaround, called `always_false<T>` which will always evaluate
to a `false` state.