🔨 Detect multi-unit encoding from max units
The encoding objects currently require the encoding to identify whether
it is multi-unit by the presence of an `is_multi_unit` bool. Although
this works, this lacks sufficient information about the encoding -- such
as the max unit-encoding size, which may be relevant for other things
such as reversing encoding patterns.

This updates the detection mechanism to make use of a metric of counting
the max code units per code-point, `max_units_per_char`. If not specified,
the encoding is assumed to not be multi-unit-encoded.