📖 Clean up README
This project hasn't been worked on in ages. Some of the links are not
relevant, some are just out-of-date.
As a first start to revitalizing this project, this does a quick pass of
doc cleanups to make it more presentable.

More work will be done provide details for getting started with the
engine