✨ Add dev-container image and configuration
It's been a goal for a long time to attempt to dockerize C++ builds, but
I simply never had the experience or the time to focus on learning how
to do this. Having now spent the last 2 years as a dev-tool engineer,
I've learned a lot about how to configure this sort of technology, and
how to use it more effectively. This includes how to now configure
cross-compilation support so that we can evaluate the behavior for
building on Linux or Windows from the container. The only one system
that can't be containerized is macOS, but that is okay since this is
being developed _on_ macOS.

This adds in a basic `dev-container.Devshell` configuration that
specifies a Dockerfile that sets up a Ubuntu-based container with
all the necessary dependencies to build the project. It also includes
a VSCode configuration that will automatically install the necessary
extensions and settings to make the experience as smooth as possible.