Add charset subset encoding traits
We have support for complex string-encoding conversions for UTF
string-views and strings, but we have no way of identifying when two
encodings are compatible or subsets of one-another. This commit
introduces a new trait to help identify subset encodings, such as
ASCII being a subset of UTF-8, or ANSI being a subset of Latin-1.

Change-Category: improvement