Fix encoding length types
The encoding traits are attempting to convert `std::ptrdiff_t`, which
are signed values, as `std::size_t` in the `ucapacity` type, which is
unsigned. This was leading to warnings about signed to unsigned
conversions.

This change explicitly casts the distance to `std::size_t` to
silence those warnings -- since the distances cannot be negative in
this context.

Change-Category: bugfix